+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.5.6 [????.??.??]
+* Support deriving `Eq`, `Ord`, and `Show` instances for data types with fields
+  of type `Int8#`, `Int16#`, `Word8#`, or `Word16#` on GHC 8.8 or later.
+
 ### 0.5.5 [2019.04.26]
 * Support `th-abstraction-0.3` or later.